refactor: view를 위한 enum 위치 이동 및 클래스명 변경
- ReservationStatus가 예약인지 대기인지 판별하는 것은 view를 위한 enum이었음
